clear up the logic
